ConfigDialog: don't warn about changes while resetting previous page
When the user makes a change on a page of the Configure KDevelop dialog,
then switches to another page without applying, the Apply Settings
dialog appears. If the user clicks the Discard button in this dialog, a
warning or two like the following are usually printed:
    kdevplatform.shell: Settings in config page
    UiPreferences(<pointer>, name = "UiConfig") changed, while
    KDevelop::PluginPreferences(<pointer>) is currently selected. This
    case is not implemented yet.

The warning is printed during the oldPage->reset() call in
ConfigDialog::checkForUnsavedChanges() and/or when control returns to
the event loop as the comment added to checkForUnsavedChanges()
explains.

Store a pointer to the page that is being reset in a new data member of
ConfigDialog. Do not print the warning when a change occurs on the
non-current page that is being reset, because KDevelop handles this
scenario correctly.