SourceFormatterSelectionEdit: block signals while adding style item
At some point since 7411a92e54fb8a70aa5b2387bb4051ea5a54f555, possibly
due to the porting to Qt 6, QListWidget started emitting the signal
itemSelectionChanged() when a new item is inserted at the top, even
though the same previously selected item remains selected. The
unexpected signal emission reverts the selected style of the current
language from the new style to the old one during the call to
SourceFormatterSelectionEditPrivate::addStyleItem() from
SourceFormatterSelectionEdit::newStyle(). The wrong value of
LanguageSettings::selectedStyle() causes an assertion failure with the
following call stack:
ASSERT: "style == styleSelectedInUi()" in file kdevelop/kdevplatform/shell/sourceformatterselectionedit.cpp, line 610
SourceFormatterSelectionEditPrivate::currentStyle()
SourceFormatterSelectionEditPrivate::updateStyleButtons()
SourceFormatterSelectionEditPrivate::updateUiForCurrentStyle()
SourceFormatterSelectionEdit::newStyle()

Start blocking signals of the style list widget before the call to
addStyleItem() in newStyle() in order to prevent the problematic
invocation of SourceFormatterSelectionEdit::styleSelectionChanged().

Also block signals during a subsequent call to QListWidget::editItem()
so as to be able to use the local variable newStyleItem that is now
declared in the scope of the list widget's QSignalBlocker. Starting
editing of an item does not emit any of the signals, which
SourceFormatterSelectionEdit connects to, so this change is
inconsequential.