Extract KDevMI::DebuggerTestBase
The extracted common base class of GdbTest and LldbTest reduces the
amount of boilerplate code in the two test classes. Previously, common
nonmember test functions had to be called from the same-name private
slots of GdbTest and LldbTest. This commit converts these common
functions into private slots of the new class DebuggerTestBase and
eliminates the corresponding trivial private slots of GdbTest and
LldbTest.

A side effect of this refactoring is that now the common test functions
run before all other test functions. This is fine because separate test
functions do not and should not depend on each other.

Move functions that depend on state (e.g. on a function-local static
constant or on KDevelop::ICore) from the namespace KDevMI::Testing into
the new class DebuggerTestBase in order to replace the function-local
static variables with data members and to reuse the KDevelop::ICore*
data member.

QSKIP() test_lldb if lldb-mi is unavailable at the beginning rather than
at the end of initTestCase() so as to avoid introducing one more virtual
function finishInitTestCase(). initTestCase() does not test much, and
the same code is also invoked by test_gdb, so there is no benefit in
running the setup code if test_lldb is to be skipped.

Introduce and use two convenience macros START_DEBUGGING_E() and
START_DEBUGGING_AND_WAIT_FOR_PAUSED_STATE_E() to avoid repeating
QVERIFY() and executePlugin() everywhere. This commit has to modify the
lines where the new macros are used in any case because the m_iface data
members and the executePlugin function arguments are replaced with the
executePlugin() member function.