Don't (x)fail test_lldb when LLDB-MI stops at a transitory breakpoint
`test_lldb testInsertAndRemoveBreakpointWhileRunning` occasionally (in
less then 10% of test runs) fails on my GNU/Linux system with an output
like the following:
QWARN  : kdevelop.plugins.debuggercommon: Received a modification of an unknown breakpoint
QWARN  : current state KDevelop::IDebugSession::PausedState waiting for KDevelop::IDebugSession::EndedState
FAIL!  : KDevMI::LLDB::LldbTest::testInsertAndRemoveBreakpointWhileRunning() Timeout before reaching state 6
   Loc: [kdevelop/plugins/debuggercommon/tests/debuggertestbase.cpp(405)]
(sometimes with "ActiveState" or "StoppedState" in place of
"PausedState")

9eb4f7864a7eaf3117703cea265013441d0c2467 removed the line
WAIT_FOR_A_WHILE(session, 3000); // give slow debugee extra time to run
before the final waiting for the ended state from this test function.
The extra waiting was clearly not needed at that time, because
WAIT_FOR_STATE() had waited for 50 seconds. A subsequent commit
52d9bab0648bf9d10bb0812195512bda6436b868 restored the older 10-second
interval when waiting for the ended state. The test failures where the
current state is active or stopped are apparently caused by the now
insufficient waiting time. Restore the WAIT_FOR_A_WHILE() call, but wait
for 2 instead of the former 3 seconds because
9eb4f7864a7eaf3117703cea265013441d0c2467 also increased the first wait
interval in this function from 1 to 2 seconds, and because when the
debug session finishes and is destroyed during the call to
WAIT_FOR_A_WHILE(), the test function fails.

`test_lldb testInsertAndRemoveBreakpointWhileRunning` expects failure on
FreeBSD because LLDB-MI always stops at the transitory breakpoint on
that OS. Now it turns out that such stopping also occurs on GNU/Linux
sometimes - when the current state is paused while the test waits for
the ended state. I have tried replacing the first wait interval of 2000
milliseconds with 1800, 2100, 2150, 2200, 2300 and 2400 milliseconds,
but the test function remains flaky in any case (the frequency of
failures depends on the wait interval).

As the test function verifies the absence of a crash, stopping at the
transitory breakpoint should not be considered a failure. Remove the
QEXPECT_FAIL() call on FreeBSD. Detect a paused state and continue
debugging on any OS instead. Also detect a possible second paused state
and fail in that unreasonable case.

The asynchronous invocation of IDebugSession::run() in the lambda
connected to the signal IDebugSession::stateChanged() always triggered
an internal debugger error on my system. Invoking via
QTimer::singleShot() with a non-zero interval usually prevented this
error. But when the error occurred with the 300 ms interval (once, after
more than 100 test runs), I decided to analyze and get rid of this
source of test delay and flakiness.

Print an error message before showing the "Internal debugger error"
message box to clarify when exactly the error has occurred. This is
useful because event processing and consequently KDevelop output
continues while the message box is displayed. Printing the error message
will also be helpful if an internal debugger error occurs in a unit test
on a CI server.

I thought that perhaps the inexplicable consistent total freezes of
test*Attach() on the FreeBSD CI server are caused by the internal
debugger error message box. Checked this by not skipping the tests, but
the added error message was absent from the test output. Therefore,
something else must be causing those mysterious freezes, and so we have
to keep skipping test*Attach() on FreeBSD.

When the debug session is active, LLDB-MI replies to the
stack-list-frames command with `stack=[{}]`. FrameListHandler::handle()
processes this reply and unconditionally accesses the field "level" of
stack[0]. As the field is absent from the empty-stack LLDB-MI reply,
KDevMI::MI::TupleValue::operator[] throws a type_error, which triggers
the internal debugger error. Detect and handle this empty-stack reply
separately.

I have also tried calling IDebugSession::run() from the lambda
synchronously. This sent the exec-continue command even earlier, before
the thread-info command, and triggered another internal debugger error.
That's because when the debug session is active, LLDB-MI replies to the
thread-info command with `threads="[]"`. The type of `threads` must be
`list`, not `const`, so this is probably a bug in LLDB-MI. The wrong
type makes the threads.size() call in
LldbFrameStackModel::handleThreadInfo() throw a type_error. Work around
this LLDB-MI bug by detecting and handling the incorrect reply
separately.

The two special LLDB-MI replies handled in this commit might (rarely) be
received by KDevelop itself, not only by unit tests. So this commit may
prevent some real internal debugger errors and consequent user bug
reports in addition to making
`test_lldb testInsertAndRemoveBreakpointWhileRunning` non-flaky.