Always select an index of MultiLevelListView after KNewStuff changes
At some point since KDevelop version 5.10.221202 or KDE Frameworks
version 5.103 or Qt version 5.15.8, a segmentation fault in
QAbstractProxyModelPrivate::emitHeaderDataChanged() began to occur
whenever the user closes the dialog "Get More Templates..." of the
dialog "Create Files from Template". The same crash also occurs when the
user uninstalls a previously installed application template, does not
install any other template, and closes the dialog "Get More Templates"
of the dialog "Create New Project".

The root cause of the regression is not clear to me, perhaps an internal
change in Qt. Fortunately, the crash can be eliminated by always
selecting an index of MultiLevelListView after its model is reset.
MultiLevelListView is not a real item view, does not connect to the
signals of its model, and therefore requires manual setting of the
current index.

Extract a new class TemplatesViewHelper to deduplicate the slot/function
handleNewStuffDialogFinished() of ProjectSelectionPage and
TemplateSelectionPagePrivate.

The replaced slot ProjectSelectionPage::handleNewStuffDialogFinished()
correctly returned early in case nothing has changed, helpfully tried to
select one of the newly installed templates, and fell back to selecting
the first template because something should always be selected in a
MultiLevelListView. But this function incorrectly assumed that if the
list KNSCore::Entry::installedFiles() is nonempty, its first element is
always present in the TemplatesModel. Turns out that calling
installedFiles() on an uninstalled entry may return a nonempty list,
possibly due to a bug in KNewStuff.

Now when the installed file of an entry is absent from the model,
TemplatesViewHelper::handleNewStuffDialogFinished() tries to select
another of the newly installed templates. And if nothing new can be
selected, it reliably falls back to selecting the first template.
Another difference from the replaced slot
ProjectSelectionPage::handleNewStuffDialogFinished() is that a warning
is now printed when there is more than the expected one installed file.

The function
TemplateSelectionPagePrivate::handleNewStuffDialogFinished() of the
"Create Files from Template" dialog always refreshed the TemplatesModel,
even if the user did not make any changes in the "Get More Templates..."
dialog. And it never selected any MultiLevelListView index. Therefore,
back when the crash did not occur, all items of the view were deselected
after closing the KNewStuff dialog. Now that the code is reused, the
behavior is better than before and the same as in the
"Create New Project" dialog.

Extract view level count and view level constants, and use them in place
of magic numbers.

Emit the signal MultiLevelListView::currentIndexChanged() at the
beginning rather than the end of MultiLevelListView::setCurrentIndex().
This way, the very likely other emission of this signal triggered by the
setting of valid indices in MultiLevelListView::setCurrentIndex()
overrides rather than is overridden by it. When an invalid index is set
as current as a fallback after closing the KNewStuff dialog of the
"Create Files from Template" dialog, the new order of emissions prevents
deselecting all templates.

BUG: 504104
FIXED-IN: 6.3.250800