Always select an index of MultiLevelListView after loading templates
Loading an invalid application or file template from a file in the
"Create New Project" or "Create Files from Template" dialog causes the
same segmentation fault in
QAbstractProxyModelPrivate::emitHeaderDataChanged() as the parent commit
eliminated for KNewStuff changes. This commit eliminates the crashes
after loading templates in the same way - by always selecting an index
of MultiLevelListView after its model is reset.

Extract TemplatesViewHelper::loadTemplatesFromFiles() to deduplicate the
slot/function loadFileClicked() of ProjectSelectionPage and
TemplateSelectionPagePrivate. Only
ProjectSelectionPage::loadFileClicked() passes the caption
"Load Template from File" to QFileDialog(). Pass the same caption in the
extracted function because it is more informative than the default
caption "Open".

The replaced slot ProjectSelectionPage::loadFileClicked() set as current
only a loaded template with 3 or more indexes in the TemplatesModel. Now
a 2-index template, also supported by ProjectSelectionPage, is set as
current too, because ProjectTemplatesViewHelper::setCurrentTemplate() is
reused instead of the custom template selection code.

The replaced function TemplateSelectionPagePrivate::loadFileClicked()
set as current the second index of a loaded template with 2 or more
indexes in the TemplatesModel. This is wrong because only templates with
3 or more indexes are supported by TemplateSelectionPage and because the
last (at least the third) index should be set as current to select the
loaded template itself rather than the first template in its Framework
category. Now the last index of a template with 3 or more indexes is set
as current because FileTemplatesViewHelper::setCurrentTemplate() is
reused instead of the custom template selection code.

If none of the loaded templates can be selected in the UI, reuse the
functions makeFirstTemplateCurrent() of ProjectSelectionPage and
TemplateSelectionPagePrivate introduced in the parent commit to select
the first template of MultiLevelListView as a fallback, and thereby
prevent the crash.

Do not refresh the model in TemplatesModel::loadTemplateFile().
loadTemplateFile() is always invoked in a loop - in the just extracted
TemplatesViewHelper::loadTemplatesFromFiles() and (via
ITemplateProvider::loadTemplate()) in TemplatePage::loadFromFile().
Refresh the model once after the loop instead as
TemplatePage::loadFromFile() already does via
ITemplateProvider::reload(). This optimization also allows to resort to
the fallback of selecting the first template only if none of the loaded
templates can be selected.