Reuse TemplatesViewHelper in TemplatePage
Reusable functions TemplatesViewHelper::handleNewStuffDialogFinished()
and TemplatesViewHelper::loadTemplatesFromFiles() perform the same work
as TemplatePage::handleNewStuffDialogFinished() and
TemplatePage::loadFromFile() respectively, and then select one of the
downloaded or loaded templates in the UI. Reuse the class
TemplatesViewHelper instead of the custom functions to deduplicate and
to add the template selection feature to the Templates page of the
Configure KDevelop dialog.

Reusing TemplatesViewHelper requires changing the interface class
ITemplateProvider to return the more concrete model type supplied by the
two existing template providers - TemplatesModel instead of
QAbstractItemModel. Reusing TemplatesViewHelper and exposing
TemplatesModel allows to remove 3 functions of ITemplateProvider that
are trivially and equivalently implemented by the two concrete template
providers.

Transfer the ownership of the model returned by ITemplateProvider from
the provider to the user of the interface to simplify the code. This
simplification is also an optimization:
1. The template provider plugins no longer store a TemplatesModel
   instance. So a TemplatesModel exists only while being used - when
   "Create New Project" or "Create Files from Template" or
   "Configure KDevelop" dialog is open.
2. TemplatesModel::refresh() is no longer called redundantly.

Eliminate class ProjectTemplatesModel, which merely encapsulates the
model type prefix and calls refresh() in the constructor. This way users
of ITemplateProvider can call refresh() themselves during initialization
as required by an ordinary TemplatesModel, which never refreshes itself.

Pass ITemplateProvider& instead of ProjectTemplatesModel* from
AppWizardPlugin through AppWizardDialog to ProjectSelectionPage in order
to also reuse ITemplateProvider::knsConfigurationFile() instead of
duplicating the string literal.

Pass ITemplateProvider& from FileTemplatesPlugin or testfiletemplates
through TemplateClassAssistant to TemplateSelectionPage in order to
reuse ITemplateProvider::createTemplatesModel() (the model type prefix)
and ITemplateProvider::knsConfigurationFile(). The test program
testfiletemplates does not create a FileTemplatesPlugin instance, so
introduce an intermediary base class of FileTemplatesPlugin -
FileTemplateProvider, which implements the interface ITemplateProvider.