TemplateSelectionPage: don't select first leaf TemplatesModel index
TemplateSelectionPage() finds the first leaf index of TemplatesModel and
sets it as current in MultiLevelListView if the project/session config
does not contain or contains an invalid last used template entry. The
indexes of TemplatesModel are not sorted, so an unpredictable template
somewhere in the middle of the view is usually selected. This
misleadingly suggests that the initially selected template has some sort
of priority over others.

Set an invalid index as current instead.
MultiLevelListView::setCurrentIndex() selects the first index of a
sorted proxy model at each tree level, so the very first template of the
view is now selected.