Don't emit MultiLevelListView::currentIndexChanged() redundantly
The emission of the signal currentIndexChanged() at the beginning of
MultiLevelListView::setCurrentIndex() is normally followed by another
emission of the same signal from
MultiLevelListViewPrivate::viewSelectionChanged(). If not, it means that
setting the current index has failed, and therefore the initial emission
was incorrect.

TemplateSelectionPage relies on this incorrect initial emission when it
sets an invalid index as current after refreshing the model makes it
empty. Move the UI-disabling code from
TemplateSelectionPagePrivate::currentTemplateChanged() into a new
function TemplateSelectionPagePrivate::currentTemplateInvalidated() and
call it after refreshing the model. Remove the no longer needed initial
emission of MultiLevelListView::currentIndexChanged() and document that
the signal is not emitted when the current index becomes invalid.

Assert that the new current index is valid instead of checking this no
longer possible condition in the slot and the function connected to
MultiLevelListView::currentIndexChanged(). Also assert that the new
current index does not have children instead of checking this impossible
even before this commit condition in the connected function.

Do not call the aforementioned connected slot from
ProjectSelectionPage() if the current MultiLevelListView index is
invalid. Such a call printed a debug message and returned early before
this commit but would trigger an assertion failure now.