Fix and improve API PluginController::hasUnresolvedDependencies()
0dc08f8377090897715c7629e07ca1b410979971 renamed PluginController's
member function checkForDependencies() to hasUnresolvedDependencies()
and documented it. However, the added documentation is
self-contradictory: "this method returns true. Otherwise, @p missing is
empty and this method returns true". Also the return value is the
inverse of what the function name implies and most of the documentation
says.

Return the list of unresolved dependencies by value instead of the bool
return value and the list output parameter. Rename the function again to
unresolvedDependencies() and adjust its documentation accordingly.

This refactoring simplifies the definition and the documentation of the
function, and does not complicate the code that uses it. The removal of
the output parameter also makes using the function less error-prone - no
need to worry whether the function clears the list argument if there are
no unresolved dependencies.