RunController: read from configs launch configurations of added type
When a plugin is unloaded and removes a configuration type,
RunController::removeConfigurationType() removes existing launch
configurations of this type. In contrast, if a plugin is loaded and adds
a configuration type after RunController::initialize(), launch
configurations of this type are not read from config. Such late loading
occurs when the user enables a plugin on the Plugins page of the
Configure KDevelop dialog.

Read launch configurations of types added while the RunController is in
the initialized state. Check RunControllerPrivate::executeMode for null
to determine whether the RunController is currently in this state. The
data member executeMode is initialized to null, becomes non-null in
RunController::initialize(), and becomes null again in
RunController::cleanup().