Revert a porting mistake in PluginControllerPrivate::canUnload()
bdc7b13521f60b5cab077efb946c11ca3ef6531b made a mistake in
PluginControllerPrivate::canUnload() while porting KDevelop from
KPluginInfo to KPluginMetaData. So now the dependencies of the plugin
that is being unloaded are compared to the same plugin's own interfaces.

Restore the more logical original behavior - compare the dependencies of
other loaded plugins to the interfaces of the plugin that is being
unloaded.