Make the Git Support plugin Global
If the user disables the GitHub Provider plugin, keeps the Git Support
plugin enabled, and a Git-managed project is opened, the Git Commit tool
view is added to all sublime areas. This occurs because of the following
circumstances:
1. PluginController::initialize() loads only Global plugins and their
   dependencies.
2. "X-KDevelop-Category" is not defined in kdevgit.json, so the Git
   Support plugin is not a Global plugin.
3. The GitHub Provider plugin is the only plugin that depends on the Git
   Support plugin. The GitHub Provider plugin is Global. So unless it is
   disabled, PluginController::initialize() loads both the GitHub
   Provider plugin and its dependency - the Git Support plugin.
4. If the Git Support plugin is not already loaded when a Git-managed
   project is opened, ProjectPrivate::loadVersionControlPlugin() called
   from Project::open() asks the PluginController to load it.
5. On KDevelop start, projects are opened after
   CorePrivate::initialize() calls UiController::loadAllAreas(). So
   ProjectPrivate::loadVersionControlPlugin() is invoked after
   UiController::loadAllAreas().
6. If the Git Support plugin is loaded during the opening of a project,
   GitPlugin() calls UiController::addToolView() for the Git Commit tool
   view after UiController::loadAllAreas() returns. In this case
   UiControllerPrivate::areasRestored is true, and therefore the Git
   Commit tool view is added to all sublime areas.

Adding a tool view to all sublime areas overrides repeated explicit
removal of the tool view by the user, which is very annoying. Set
"X-KDevelop-Category" to "Global" in kdevgit.json to load the Git
Support plugin earlier and no longer force the Git Commit tool view on
the user. All other plugins that add tool views are already Global
except for:
* the Patch Review plugin, which avoids adding its tool view by passing
  IUiController::None to UiController::addToolView();
* the Output View plugin, which adds output tool views only on demand,
  when output is generated. Adding the output tool views to all sublime
  areas is useful. For example, if a debug session ends, the debug
  output can still be examined in the Code sublime area.

2dff0f24356aa615c1f3b725ce183fabaf916e6d removed the
"X-KDevelop-Category": "Global" entry from kdevgit.json. Later the entry
was added to .json files of all plugins, but
df99304843fbdb593398e3d7d912e15cbd7c56b9 reverted most of the additions,
including in kdevgit.json, on the next day because they broke a few unit
tests. Re-adding the entry to kdevgit.json in this commit does not break
any unit tests.