TemplatePage: always update enabled state of Clone Template button
When the TemplatesModel is refreshed (reset), but no new template is
downloaded or loaded or selecting a new template fails, the enabled
state of the Clone Template button is not updated. So if a template
archive was selected before the template downloading or loading, the
Clone Template button is enabled even though an archive extraction (the
cloning) is impossible. Clicking the button in this scenario simply
disables it.

Disable the Clone Template button when a new template is not selected.