ProjectSelectionPage: don't emit valid() redundantly
The signal invalid() can be emitted after the signal valid() in
ProjectSelectionPage::validateData() since
263040e29fd5d4ec72504cb6c1d76ef26cf8b7a1. The emission of valid() is
redundant when this occurs. Optimize by hiding the error message and
emitting valid() only after all error checks and early returns.

The aforementioned commit also adjusted a comment, but the comment
change was reverted (probably during a careless merge). Restore the
adjusted version of the comment that explains the current implementation
accurately.