Reselect previously selected template as a fallback
When a newly installed or loaded template cannot be selected, either the
very first template of the view becomes selected or, in the case of
TemplatePage, everything is deselected. Reselect the previously selected
template instead. The previously selected template is more likely to be
used and is more difficult to find than the very first template. Not
changing the selection also better conveys that a new template could not
be selected.