TemplatePage: extract and simplify templateArchiveFileName()
Call QModelIndex::data() instead of QAbstractItemModel::data() to
simplify the definition of the extracted function. The model of a
QTreeView's index is either null if the index is invalid or equals the
tree view's model. And QModelIndex::data() returns the model's data()
for the index or {} if the index is invalid.