test_duchain: triple the timeout in ThreadList::join()
TestDUChain::testLockForRead() has failed on the FreeBSD CI server with
the following output:
FAIL!  : 'threads.join(1000)' returned FALSE. ()
   Loc: [kdevelop/kdevplatform/language/duchain/tests/test_duchain.cpp(744)]
QFATAL : QThread: Destroyed while thread is still running
FAIL!  : TestDUChain::testLockForRead() Received a fatal error.
   Loc: [qtbase-everywhere-src-6.8.2/src/corelib/thread/qthread.cpp(470)]

I suspect that the FreeBSD CI server was very busy with some other work
while running KDevelop tests because
TestBackgroundparser::testParseOrdering_foregroundThread() also failed
due to slower than usual execution in the same CI job.

Triple the timeout in ThreadList::join() so as to wait longer for the
threads to finish. Also call quit() on all threads first, then wait for
all threads to finish to speed up the test. These changes should make
the test non-flaky.

Replace Q_ASSERT() with QVERIFY() because Q_ASSERT should not be used in
tests. See
https://doc.qt.io/qt-6/qttest-best-practices-qdoc.html#avoid-q-assert

QVERIFY() that the threads finished in time inside ThreadList::join()
instead of returning a bool and verifying it in the calling functions.
ThreadList::join() is called only at the end of 3 test functions, so
there is no ambiguity where the failure has occurred (one cannot tell
which of the 10 threads could not be joined in time, but it was the same
before this commit too). If ThreadList::join() is called multiple times
in some test function in the future, RETURN_IF_TEST_FAILED() should be
used instead of returning and verifying a bool.