Extract SimpleDataToolViewFactory
The two class templates named DebuggerToolFactory in the namespaces
KDevelop and KDevMI are equivalent except that the former stores a raw
DebugController pointer and the latter - a QPointer<Plugin>. Extract a
common class template SimpleDataToolViewFactory into util because it is
general enough to be reused elsewhere.

Inherit the new class from a base class template and that from a base
class (as one of the deduplicated DebuggerToolFactory class templates
did) to reduce template code bloat.