Don't register duplicate debugger plugin tool views
Since 63e47f1d09dfa793f34bca256d1e479d849fbd11 or more specifically
since its sub-commit
https://phabricator.kde.org/R32:2d142a20c20da883fb9eed0707b29dcb47daa803
debugger plugin tool views are registered in the constructors and
unregistered in the destructors of DebugSession objects rather than in
the constructors and unload() member functions respectively of the
plugin instances themselves. This was done to avoid displaying both GDB
and LLDB tool views during every debug session.

But the current simple tool view (un)registration implementation does
not support starting a new debug session of the same debugger plugin
before the previous one finishes. In this case setupToolViews() is
called for the second time before unloadToolViews() is ever called. As a
result, duplicate tool views are registered and the data members of the
debugger plugin point to the last-created tool view factories. When the
old, replaced session is destroyed, the newer tool views are
unregistered and the factory data member(s) of the debugger plugin are
set to nullptr. The old tool views are forgotten about, and so remain
even after the second debug session finishes - in all sublime areas.

Another bug in the current implementation is that the debugger plugin
tool views are never unregistered if a debugger plugin is unloaded
before its debug session ends.

Introduce dedicated ToolViewFactoryHolder classes that wrap
(un)registration of the tool views in their constructors and destructors
(RAII). Let a debug session keep a ToolViewFactoryHolder only while it
is the current session. When another debug session replaces the current
session of the same debugger plugin, reuse the existing tool views.

The new implementation takes care to preserve positions and visibility
of tool views across debug sessions. But if a new debug session replaces
the current session of a different debugger plugin, tool views are added
and removed while the Debug sublime area is the current area. In this
case the custom positions and visibility of tool views may be lost. This
will have to be fixed in Sublime and the UiController: (un)registering a
tool view should not affect whether or not it is "desired" by sublime
areas, and QMainWindow::restoreDockWidget() should be used to restore
the visibility and state of a tool view added after the main window
settings are loaded.