Disconnect from no longer current debug session
When the user starts a new debug session, an IDebugSession object is
created and passed to DebugController::addSession(), which connects the
DebugController to the new IDebugSession instance and emits a signal
currentSessionChanged(). Debugger tool views connect to the signals and
slots of the newly current IDebugSession instance in their slots
connected to the signal currentSessionChanged().

When the user starts a new debug session while the previous debug
session is running, the DebugController itself and all debugger tool
views fail to disconnect from the signals and slots of the replaced, no
longer current, IDebugSession instance. The signals of the old
IDebugSession instance are erroneously handled while it finishes.

Due to a particularly slow start of LLDB debugging on my system, I could
reproduce bugs with the following steps:
1. Start an LLDB debug session.
2. Immediately start a GDB debug session that stops at a breakpoint very
   soon after starting.
If the GDB session pauses before the LLDB session ends, the Frame Stack
tool view ends up disabled (because
FramestackWidget::sessionStateChanged() handles the ended state of the
old session after the paused state of the new session) and the execution
mark is missing (because DebugController::clearExecutionPoint() for the
old session is invoked after DebugController::showStepInSource() for the
new session).

Add a second parameter named previousSession to the signal
currentSessionChanged() and use it in the slots of the debugger tool
views to disconnect from the no longer current session. Also disconnect
the DebugController itself from some signals of the previous session.
This eliminates both the reproduced bugs and other potential bugs.

Use the new previousSession parameter in
MIDebugSession::currentSessionChanged() to make the
stop-being-current-session check more exact, and thus (slightly)
optimize.

Rename DebuggerConsoleView::handleSessionChanged() to
currentSessionChanged() for consistency with other debugger tool views.

Convert the protected or private slots currentSessionChanged() into
private member functions because they are connected only via the
function pointer-based syntax and are used only by their own respective
classes.