Finish debug job whenever its debug session finishes
MIDebugSession emits IDebugSession::finished() only when the debugger
stops - in MIDebugSession::handleDebuggerStateChange(). But this signal
is not emitted if the session ends before the debugger starts - in
MIDebugSession::stopDebugger(). Only MIDebugJobBase connects to the
finished() signal - to finish the debug job associated with the session.

As a result, the following steps cause a segmentation fault when
MIDebugJob::start() dereferences the job's null m_session:
1. Start debugging with a build dependency.
2. While the not-yet-started debug session waits for the build
   dependency job to finish, start another debug job without a build
   dependency.
3. Agree to abort the current debug session in the warning dialog that
   appears.
4. Wait for the build job to finish and witness a segmentation fault
   when the instance of ExecuteCompositeJob created in
   makeJobWithDependency() starts its next subjob - MIDebugJob.

Finish a debug job when its session emits the more widely used and
reliable alternative signal IDebugSession::stateChanged() with the state
argument equal to IDebugSession::EndedState instead of when the
problematic signal IDebugSession::finished() is emitted.

A debug job calls MIDebugSession::stopDebugger() in case of a launch
configuration error or when it is killed or destroyed while unfinished.
stopDebugger() cannot possibly emit the finished() signal synchronously,
so the about-to-be-destroyed debug job had no need to disconnect from
that signal. However, stopDebugger() does emit the signal
stateChanged(EndedState) synchronously in case the debugger is not
started. The connection added in this commit then finishes the job at an
unexpected time, possibly before it is even stared, which can lead to a
segmentation fault on KDevelop exit because of a dangling job pointer
registered with the RunController. Disconnect from the stateChanged()
signal before stopping the debugger to prevent such issues.

84ffb556e22651c8508beb6a40edf10110f8ae8a changed the type of
MIDebugJobBase::m_session from raw pointer to QPointer in order to
prevent a crash when the session was destroyed but the job did not
finish before being destroyed. The crash could occur only because the
finished() signal was not emitted. This commit connects to a reliable
signal instead, so revert the no longer needed QPointer change.

Remove the no longer connected-to signal IDebugSession::finished(). This
signal is/was emitted by the debug sessions and connected to by the
debug jobs of the external debugger plugins kdev-python and kdev-xdebug.
The external plugins should either connect to the stateChanged() signal
instead as well - as kdev-python already does since
https://commits.kde.org/kdev-python/3c2e91a7f6351df166fe4ce560523f962bd48f3b
- or declare the finished() signal in their
custom debug session classes to replace the removed signal.
Seeing as kdev-xdebug has not been ported to KF6 yet, the following
issue tracks the necessary adjustment in that plugin:
https://invent.kde.org/kdevelop/kdev-xdebug/-/issues/4

Set the session state to EndedState in the lambda within
TestDebugSession::initializeDebugger(), which causes MIDebugSession to
emit the stateChanged() signal, instead of emitting the removed
finished() signal. Also set the debugger state in the same lambda to the
not-started debugger flags in order to prevent ~MIDebugSession() from
calling stopDebugger(), which invokes a pure virtual method
MIDebugSession::frameStackModel() that aborts test_midbus.

The pure virtual method call to frameStackModel() became possible since
774b1846680f2c9d0fc966779cb3fba52c35201a moved the stopDebugger() call
from the destructor of GDB::DebugSession to the destructor of its new
base class (which was eventually renamed to MIDebugSession). The
current implementation may have to be revisited at some point. For now,
when MIDebugSession ends, the debugger flag s_dbgNotStarted is always
on, and a MIDebugSession is destroyed (by DebugController) only when it
ends. Therefore ~MIDebugSession() never calls stopDebugger() in
practice.