Kill dependency job when the dependent job finishes
When a debug session replaces a previous debug session with a still
running nonconflicting dependency job, the no longer relevant dependency
(build) job keeps running and wasting resources in the background. The
same waste occurs when executing a program with a modified dependency
kills the previous execution job (NativeAppJob) of the same name, whose
dependency (build) job is still running.

Kill the obsolete dependency job when the dependent job finishes to
optimize.

Such a killing can be triggered unintentionally when a dependency build
job fails or is killed by a replacement build job (in
NinjaBuilder::runNinja() or MakeBuilder::runMake()). In such cases
deleteLater() is called in KJob::finishJob() on the composite job a bit
earlier than on the dependency job. So the dependent job, which is an
unstarted subjob and a child object of the composite job, is destroyed
and therefore finishes just before the dependency job is destroyed (but
after the dependency job finishes). So the lambda in
makeJobWithDependency() is invoked, prints a debug message and calls
kill() on the already finished dependency job. KJob::kill() returns true
immediately if the job is already finished, so the only ill effect of
the unintended lambda invocation is the misleading debug output.
Differentiating between the intentional and the unintentional invocation
of the lambda is difficult because KJob::isFinished() is protected
rather than public, so let us live with this imperfection.

The inline helper function makeJobWithDependency() can be called even
when only one of the two plugins that implement the interface
org.kdevelop.IExecutePlugin - kdevexecute or kdevexecuteplasmoid - is
loaded. Therefore makeJobWithDependency() has no associated logging
category, so use qDebug() rather than qCDebug() in it.